rel is internal pkg, use the lazy one
