+++ actors/util/rleplus/bits_test.go
@@ -1,4 +1,4 @@
-package lazy
+package rleplus
 
 import (
 	"testing"
