+++ actors/util/rleplus/rleplus_reader.go
@@ -1,4 +1,4 @@
-package lazy
+package rleplus
 
 import (
 	"github.com/multiformats/go-varint"
