+++ actors/util/rleplus/rleplus_writer.go
@@ -1,4 +1,4 @@
-package lazy
+package rleplus
 
 import (
 	"encoding/binary"
