+++ actors/util/rleplus/runs.go
@@ -1,4 +1,4 @@
-package lazy
+package rleplus
 
 import (
 	"math"
