+++ actors/util/rleplus/runs_test.go
@@ -1,4 +1,4 @@
-package lazy
+package rleplus
 
 import (
 	"math/rand"
